refactor(ai): rename transaction/verb API to plugin, strip composition + clientTransaction
