+++ packages/ai/src/plugin.ts
@@ -1,0 +1,19 @@
+export {
+  chatPlugin,
+  definePlugin,
+  generationPlugin,
+} from './activities/plugin/index'
+export type {
+  AnyPlugin,
+  ChatPlugin,
+  ChatPluginCallback,
+  ChatPluginRequest,
+  ChatPluginReturn,
+  GenerationPlugin,
+  GenerationPluginExecute,
+  GenerationPluginOptions,
+  GenerationPluginRequest,
+  PluginConfig,
+  PluginDefinition,
+  PluginKind,
+} from './activities/plugin/index'
